requirements.txt
aniso8601==7.0.0
asn1crypto==1.2.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.1.0
promise==2.2.1
psycopg2==2.8.4
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.2
whitenoise==4.1.4
win-inet-pton==1.1.0
wincertstore==0.2
